# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controller.Message:
    properties:
      concertId:
        type: string
      isFirstMessage:
        type: boolean
      position:
        type: integer
      status:
        type: string
    type: object
...
  models.Artist:
    properties:
      concerts:
        items:
          $ref: '#/definitions/models.Concert'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      interest:
        $ref: '#/definitions/models.Interest'
      interestId:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Category:
    properties:
      concertCategories:
        items:
          $ref: '#/definitions/models.ConcertCategory'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Concert:
    properties:
      artist:
        $ref: '#/definitions/models.Artist'
      artistId:
        type: string
      concertCategories:
        items:
          $ref: '#/definitions/models.ConcertCategory'
        type: array
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        description: gorm.Model
        type: string
      image:
        type: string
      interests:
        items:
          $ref: '#/definitions/models.Interest'
        type: array
      location:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/models.Organization'
      organizationId:
        type: string
      updatedAt:
        type: string
    type: object
  models.ConcertCategory:
    properties:
      availableTickets:
        type: integer
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      concert:
        allOf:
        - $ref: '#/definitions/models.Concert'
        description: Tickets          []Ticket `gorm:"-"`
      concertId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      price:
        type: number
      soldTickets:
        type: integer
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      updatedAt:
        type: string
    type: object
  models.Conversation:
    properties:
      buyer:
        $ref: '#/definitions/models.User'
      buyerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      price:
        type: number
      seller:
        $ref: '#/definitions/models.User'
      sellerId:
        type: string
      ticketListing:
        $ref: '#/definitions/models.TicketListing'
      ticketListingId:
        type: string
      updatedAt:
        type: string
    type: object
  models.Interest:
    properties:
      concerts:
        items:
          $ref: '#/definitions/models.Concert'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Message:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorId:
        type: string
      content:
        type: string
      conversation:
        $ref: '#/definitions/models.Conversation'
      conversationId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      readed:
        type: boolean
      sentAt:
        type: string
      updatedAt:
        type: string
    type: object
  models.Organization:
    properties:
      concerts:
        items:
          $ref: '#/definitions/models.Concert'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        description: gorm.Model
        type: string
      image:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Sale:
    properties:
      buyer:
        $ref: '#/definitions/models.User'
      buyerId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      finalPrice:
        type: number
      id:
        description: gorm.Model
        type: string
      seller:
        $ref: '#/definitions/models.User'
      sellerId:
        type: string
      ticketListingId:
        type: string
      ticketSold:
        $ref: '#/definitions/models.TicketListing'
      updatedAt:
        type: string
    type: object
  models.Ticket:
    properties:
      concertCategory:
        $ref: '#/definitions/models.ConcertCategory'
      concertCategoryId:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: gorm.Model
        type: string
      maxPrice:
        type: number
      ticketListings:
        items:
          $ref: '#/definitions/models.TicketListing'
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userId:
        type: string
    type: object
  models.TicketListing:
    properties:
      conversations:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        description: |-
          gorm.Model
          ID            uuid.UUID       `gorm:"unique;type:uuid;primaryKey;default:uuid_generate_v4()" json:"id"`
          Price         float64         `gorm:"not null" json:"price"`
          Status        string          `gorm:"not null" json:"status"`
          CreatedAt     time.Time       `json:"createdAt"`
          UpdatedAt     time.Time       `json:"updatedAt"`
          DeletedAt     *time.Time      `gorm:"index"`
          TicketId      uuid.UUID       `gorm:"not null" json:"ticketId"`
        type: string
      price:
        type: number
      sale:
        $ref: '#/definitions/models.Sale'
      status:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticketId:
        type: string
      updatedAt:
        type: string
    type: object
  models.User:
    properties:
      conversationsAsBuyer:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      conversationsAsSeller:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        description: |-
          l'email de l'utilisateur

          required: true
          example: john@doe.com
        type: string
      firstname:
        description: |-
          le prénom de l'utilisateur

          required: true
          example: John
        type: string
      id:
        description: |-
          gorm.Model
          l'ID de l'utilisateur

          required: true
          example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      image:
        type: string
      interests:
        items:
          $ref: '#/definitions/models.Interest'
        type: array
      lastConnexion:
        type: string
      lastname:
        description: |-
          le nom de l'utilisateur

          required: true
          example: Doe
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      organization:
        $ref: '#/definitions/models.Organization'
      organizationId:
        type: string
      resetCodeExpiration:
        type: string
      role:
        type: string
      salesAsBuyer:
        items:
          $ref: '#/definitions/models.Sale'
        type: array
      salesAsSeller:
        items:
          $ref: '#/definitions/models.Sale'
        type: array
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the swagger documentation for the Weezemaster API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Weezemaster API
  version: "1.0"
paths:
...
  /ticketlistings:
    get:
      consumes:
      - application/json
      description: Get all ticket listings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TicketListing'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all ticket listings
      tags:
      - Ticket listing
    post:
      consumes:
      - application/json
      description: Create ticket listings
      parameters:
      - description: Ticket ID
        in: body
        name: ticketId
        required: true
        schema:
          type: string
      - description: Price
        in: body
        name: price
        required: true
        schema:
          type: number
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TicketListing'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create ticket listings
      tags:
      - Ticket listing
  /ticketlistings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete ticket listing
      parameters:
      - description: TicketListing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete ticket listing
      tags:
      - Ticket listing
    get:
      consumes:
      - application/json
      description: Get ticket listing by ID
      parameters:
      - description: TicketListing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TicketListing'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get ticket listing by ID
      tags:
      - Ticket listing
    patch:
      consumes:
      - application/json
      description: Update ticket listing
      parameters:
      - description: TicketListing ID
        in: path
        name: id
        required: true
        type: string
      - description: Price
        in: body
        name: price
        schema:
          type: number
      - description: Status
        in: body
        name: status
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TicketListing'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update ticket listing
      tags:
      - Ticket listing
  /ticketlistings/concert/{id}:
    get:
      consumes:
      - application/json
      description: Get ticket listing by concert ID
      parameters:
      - description: Concert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TicketListing'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get ticket listing by concert ID
      tags:
      - Ticket listing
...
  /ws-queue:
    get:
      description: Gère les connexions WebSocket pour la file d'attente des concerts
      operationId: handle-websocket-queue
      parameters:
      - description: ID du concert
        format: uuid
        in: query
        name: concertId
        required: true
        type: string
      - description: ID de l'utilisateur
        format: uuid
        in: query
        name: userId
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            $ref: '#/definitions/controller.Message'
      summary: Gère les connexions WebSocket pour la file d'attente des concerts
      tags:
      - WebSockets
securityDefinitions:
  Bearer:
    description: 'Provide your JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"